separate getOptimalSplit
